requirements.txt
anthropic>=0.40.0
flask>=3.0.0
APScheduler>=3.10.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
bcrypt>=4.1.0
redis>=5.0.0
boto3>=1.34.0
gunicorn>=21.2.0
pytest>=8.0.0
pytest-cov>=4.1.0
testing.postgresql>=1.3.0
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0
flask-limiter>=3.5.0
flask-smorest>=0.44.0
